requirements.txt
delta-spark==4.0.0rc1
googleapis-common-protos==1.65.0
grpcio==1.66.2
grpcio-status==1.66.2
importlib_metadata==8.5.0
numpy==2.0.2
pandas==2.2.3
protobuf==5.28.2
py4j==0.10.9.7
pyarrow==17.0.0
pyspark==4.0.0.dev1
python-dateutil==2.9.0.post0
pytz==2024.2
six==1.16.0
tzdata==2024.2
zipp==3.20.2

